+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 - OTU column in asserted distribution on Browse OTU [#1846]
 - Radial annotator on Uniquify people task
 - History title on Browse nomenclature
+- otu_ids param on Image matrix task
 
 ### Fixed
 - CVT view helper bug with predicates